requirements.txt
# Core Dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
biopython>=1.83

# Protein Language Models
transformers>=4.30.0
sentencepiece>=0.1.99

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0

# Optional: For Graph Neural Networks
# torch-geometric>=2.3.0
# torch-scatter>=2.1.0
# torch-sparse>=0.6.0

# Development
pytest>=7.4.0
black>=23.0.0
